+++ test/test_registry.rb
@@ -34,6 +34,7 @@ class RegistryTest < Minitest::Test
                                                       :ctestLog,
                                                       :dhrystone,
                                                       :lldMap,
+                                                      :scimark2,
                                                       :sizeBerkeleyStdout].sort
   end
 
@@ -56,6 +57,10 @@ class RegistryTest < Minitest::Test
 
   def test_lld_parser_entry
     template(:lldMap, BuildLogParser::LLDParser)
+  end
+
+  def test_scimark2_parser_entry
+    template(:scimark2, BuildLogParser::Scimark2Parser)
   end
 
   def test_size_parser_entry