+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.4.0 (in development)
 
+* Updated command line parser library to new name.
 * [ctest/log] Relaxed the command syntax to allow all possible characters.
 
 ## v0.3.1 (11.05.2018)