+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## v0.4.0 (in development)
 
+* Updated command line parser library to new name.
 * [ctest/log] Relaxed the command syntax to allow all possible characters.
+* Added support for parsing *scimark2* command line output.
 
 ## v0.3.1 (11.05.2018)