+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Updated command line parser library to new name.
 * [ctest/log] Relaxed the command syntax to allow all possible characters.
 * Added support for parsing *scimark2* command line output.
+* Added support for parsing lmbench *lat_mem_rd* command line output.
 
 ## v0.3.1 (11.05.2018)