+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * [ctest/log] Relaxed the command syntax to allow all possible characters.
 * Added support for parsing *scimark2* command line output.
 * Added support for parsing lmbench *lat_mem_rd* command line output.
+* Added support for parsing *gem5* generated `stats.txt` files.
 
 ## v0.3.1 (11.05.2018)